Add new file for curl snippets - headers json
Who woulda thunk it, curl is one hell of a useful program.
Starting this file with an initial commit of a oneliner to
remind me what headers to use if you are communicating with a REST api
Specifically it includes an example of what headers to use in order to
get json back.