Add a few minor notes on keycloak, nuxeo, org, oc
This commit is fairly small, one minor addition to each cheatsheet per
the topics outlined in the title. Reflects ongoing work around
keycloak, nuxeo and openshift. Orgmode is being used to document todo
lists for these projects, so the addition just explains the binding of
(org-cycle) to TAB in emacs while in evil mode. (The bindings default to
vim folding/unfolding keys).

Nuxeo, keycloak and openshift are basically what we are using over the
summer to try to stablize an oauth2 workflow. So notes are mostly
related to config of those various programs.